requirements.txt
aiolimiter==1.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
colorama==0.4.6
cryptography==43.0.3
decorator==5.1.1
distro==1.9.0
Django==5.1.2
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
openai==1.53.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tornado==6.4.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
